Do not try to contact remote if none is provided
This commit improves support for setups where no remote
audit repository has been set by adding a PushEnabled()
method to the repo class to check if a remote has been set
in the config.

All actions that depend on a remote repository now check
this flag (by directly looking at the Config object for
performance).

For consistency with other methods in the class, methods
like Pull() and Push() do nothing instead of dieing when
no remote is set.